+++ firmware/hdl/payload/gtl_data_mapping.vhd
@@ -2,6 +2,9 @@
 -- Data mapping for GTL
 
 -- Version-history:
+-- HB 2023-03-01: updated for ZDC.
+-- HB 2023-02-09: updated comments.
+-- HB 2022-10-10: inserted ZDC.
 -- HB 2022-09-05: cleaned up.
 -- HB 2021-06-01: first design.
 
@@ -89,4 +92,14 @@ begin
 
     gtl_data.ext_cond <= lhc_data.external_conditions(NR_EXTERNAL_CONDITIONS-1 downto 0);
 
+-- ****************************************************************************************
+
+-- HB 2023-03-01: ZDC data (10G) on link 11 (after conversion from 5G to 10G).
+    gtl_data.zdc(0) <= lhc_data.zdc10g_0(EN_MINUS_BIT_HIGH downto EN_MINUS_BIT_LOW);
+    gtl_data.zdc(1) <= lhc_data.zdc10g_1(EN_PLUS_BIT_HIGH downto EN_PLUS_BIT_LOW);
+    gtl_data.zdc(2) <= "0000000000"; -- not defined for ZDC, but 6 frames forseen in "lhc_data_pkg.vhd"
+    gtl_data.zdc(3) <= "0000000000";
+    gtl_data.zdc(4) <= "0000000000";
+    gtl_data.zdc(5) <= "0000000000";
+    
 end architecture rtl;