+++ scripts/mp7patch.py
@@ -8,38 +8,46 @@ import toolbox as tb
 def replace_area_constraints(filename):
     content = tb.read_file(filename)
 
-    expr_forloop = re.compile(r"(for\s+{\s*set\s*i\s*0\s*}\s+{\s*\$i\s*<\s*)(\d+)(\s*}\s+{\s*incr\s+i\s*})")
-    expr_cells = re.compile(r"(add_cells_to_pblock\s+\[\s*get_pblocks\s+payload_)(\d+)(\s*]\s*\[get_cells\s+(?:-\w+\s+)?datapath/rgen\[)(\d+)(]\.region/pgen\.\*])")
+    #expr_forloop = re.compile(r"(for\s+{\s*set\s*i\s*0\s*}\s+{\s*\$i\s*<\s*)(\d+)(\s*}\s+{\s*incr\s+i\s*})")
+    #expr_cells = re.compile(r"(add_cells_to_pblock\s+\[\s*get_pblocks\s+payload_)(\d+)(\s*]\s*\[get_cells\s+(?:-\w+\s+)?datapath/rgen\[)(\d+)(]\.region/pgen\.\*])")
 
-    content, count = expr_forloop.subn(r"\g<1>7\g<3>", content)
-    if count != 1:
-        raise RuntimeError("Could not replace the for-loop value.")
+    #content, count = expr_forloop.subn(r"\g<1>8\g<3>", content)
+    #if count != 1:
+        #raise RuntimeError("Could not replace the for-loop value.")
 
-    content, count = expr_cells.subn(r"\g<1>6\g<3>6\g<5>", content)
-    if count != 1:
-        raise RuntimeError("Could not replace add_cells_to_pblock line.")
+    #content, count = expr_cells.subn(r"\g<1>7\g<3>7\g<5>", content)
+    #if count != 1:
+        #raise RuntimeError("Could not replace add_cells_to_pblock line.")
+    
+    text = "resize_pblock [get_pblocks payload] -add {SLICE_X30Y0:SLICE_X191Y449}"
+    subs = "resize_pblock [get_pblocks payload] -add {SLICE_X0Y0:SLICE_X191Y449}"
+    flags = 0
+    
+    with open( filename, "r+" ) as file:
+        fileContents = file.read()
+        textPattern = re.compile( re.escape( text ), flags )
+        fileContents = textPattern.sub( subs, fileContents )
+        file.seek( 0 )
+        file.truncate()
+        file.write( fileContents )
+        
+#def replace_brd_decl(filename):
+    #content = tb.read_file(filename)
 
-    with open(filename, "w") as fp:
-        fp.write(content)
-        logging.info("Successfully patched area_constraints file '{}'".format(filename))
+    #expr_nregion = re.compile(r"(constant\s+N_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
+    #expr_crossregion = re.compile(r"(constant\s+CROSS_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
 
-def replace_brd_decl(filename):
-    content = tb.read_file(filename)
+    #content, count = expr_nregion.subn(r"\g<1>8\g<3>", content)
+    #if count != 1:
+        #raise RuntimeError("Could not replace the N_REGION value.")
 
-    expr_nregion = re.compile(r"(constant\s+N_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
-    expr_crossregion = re.compile(r"(constant\s+CROSS_REGION\s*:\s*integer\s*:=\s*)(\d+)(\s*;)")
+    #content, count = expr_crossregion.subn(r"\g<1>7\g<3>", content)
+    #if count != 1:
+        #raise RuntimeError("Could not replace the CROSS_REGION value.")
 
-    content, count = expr_nregion.subn(r"\g<1>7\g<3>", content)
-    if count != 1:
-        raise RuntimeError("Could not replace the N_REGION value.")
-
-    content, count = expr_crossregion.subn(r"\g<1>6\g<3>", content)
-    if count != 1:
-        raise RuntimeError("Could not replace the CROSS_REGION value.")
-
-    with open(filename, "w") as fp:
-        fp.write(content)
-        logging.info("Successfully patched brd_decl file '{}'".format(filename))
+    #with open(filename, "w") as fp:
+        #fp.write(content)
+        #logging.info("Successfully patched brd_decl file '{}'".format(filename))
 
 def insert_l1a_ttc(filename):
     content = tb.read_file(filename)
@@ -80,14 +88,14 @@ def patch_all(projectpath):
     """Batch patch all firmware files."""
     root_path = os.path.abspath(projectpath)
 
-    brd_decl_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7_brd_decl.vhd'
+    #brd_decl_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7_brd_decl.vhd'
     area_constraints_path = 'boards/mp7/base_fw/common/firmware/ucf/area_constraints.tcl'
     mp7xe_690_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7xe_690.vhd'
     #script_writer_path = 'scripts/firmware/dep_tree/VivadoScriptWriter.py'
 
     #append_vivado_rules(os.path.join(root_path, script_writer_path))
     replace_area_constraints(os.path.join(root_path, area_constraints_path))
-    replace_brd_decl(os.path.join(root_path, brd_decl_path))
+    #replace_brd_decl(os.path.join(root_path, brd_decl_path))
     insert_l1a_ttc(os.path.join(root_path, mp7xe_690_path))
 
 def parse_args():