+++ scripts/run_simulation_questa.py
@@ -65,6 +65,7 @@ vhdl_snippets_names = [
     'ugt_constants'
 ]
 
+DO_FILE_TMP = 'gtl_fdl_wrapper_tmp.do'
 DO_FILE = 'gtl_fdl_wrapper.do'
 TB_FILE_TPL = os.path.join('testbench', 'templates', 'gtl_fdl_wrapper_tb_tpl.vhd')
 TB_FILE = os.path.join('testbench', 'gtl_fdl_wrapper_tb.vhd')
@@ -216,7 +217,7 @@ class Module(object):
     def make_files(self, sim_dir, view_wave, mp7_tag, menu_path, ipb_fw_dir):  # makes files for simulation
         render_template(
             os.path.join(sim_dir, DO_FILE_TPL),
-            os.path.join(self.path, DO_FILE),
+            os.path.join(self.path, DO_FILE_TMP),
             {
                 '{{MP7_TAG}}': mp7_tag,
                 '{{VIEW_WAVE}}': format(view_wave),
@@ -259,6 +260,25 @@ class Module(object):
             os.path.join(uGTalgosPath, 'hdl', 'payload', 'gtl_module_tpl.vhd'),
             os.path.join(self.path, 'vhdl', 'gtl_module.vhd'),
             replace_map
+        )
+
+        # Create 'anomaly_detection.txt' from 'anomaly_detection.dep'
+        adt_dep_file = os.path.join(uGTalgosPath, 'cfg', 'anomaly_detection.dep')
+        adt_repl = os.path.join(uGTalgosPath, 'cfg', 'anomaly_detection.txt')
+        
+        with open(adt_dep_file) as fp:
+            adt_vhd = fp.read()
+        adt_vhd = adt_vhd.replace('src ', 'vcom -93 -work work $HDL_DIR/')
+        with open(adt_repl, 'w') as fp:
+            fp.write(adt_vhd)
+                        
+        # Insert content of 'anomaly_detection.txt' into DO_FILE
+        render_template(
+            os.path.join(self.path, DO_FILE_TMP),
+            os.path.join(self.path, DO_FILE),
+            {
+                '{{adt_vhd}}': read_file(os.path.join(uGTalgosPath, 'cfg', 'anomaly_detection.txt')),
+            }
         )
 
 def download_file_from_url(url, filename):
@@ -435,9 +455,16 @@ def run_simulation_questa(a_mp7_url, a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir
 
     algos_sim = {}
     algos_tv = {}
+    error_jsonf = {}
+    for i in range(menu.n_modules):
+        error_jsonf[i] = False
 
     for module in modules:  # steps through all modules and makes a list with trigger count and module
         jsonf = json.load(open(module.results_json))
+        errors_jsonf = jsonf['errors']
+        for err in errors_jsonf:
+            if err != "":
+                error_jsonf[module._id] = True
         counts = jsonf['counts']
         for count in counts:
             index = count['algo_index']
@@ -498,6 +525,15 @@ def run_simulation_questa(a_mp7_url, a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir
 
     sum_log.info("|-----|-----|------------------------------------------------------------------|--------|--------|--------|")
 
+    json_err_msg = True
+    for i in range(len(error_jsonf)):
+        if error_jsonf[i]:
+            if json_err_msg:
+                sum_log.info("\033[1;31m ERROR: mismatches of algos or finor @ certain bx-nr in: \033[0m ".format(i))
+                json_err_msg = False
+            json_file = os.path.join(base_dir, 'module_{}', 'results_module_{}.json').format(i, i)
+            sum_log.info("\033[1;31m {} \033[0m ".format(json_file))
+
     trigger_liste = trigger_list(testvector_filepath)  # gets a list: index is algorithm index and content is the trigger count in the testvector file
 
     # prints bits which are present in the testvector but have no corresponding algo in the menu
@@ -536,16 +572,22 @@ def run_simulation_questa(a_mp7_url, a_mp7_tag, a_menu, a_url_menu, a_ipb_fw_dir
 
     print("")
 
-    if success:
-        logging.info(success_green)
+    if not json_err_msg or not success:
+        logging.info(failed_red)
     else:
-        logging.error(failed_red)
+        logging.error(success_green)
 
     logging.info("===========================================================================")
     logging.info("removed temporary directory ('temp_dir') ...")
     if os.path.exists(os.path.join(sim_dir, "temp_dir")):
         shutil.rmtree(os.path.join(sim_dir, "temp_dir"))
 
+    # remove 'anomaly_detection.txt'
+    cfg_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), '..', 'firmware', 'cfg')
+    adt_txt = os.path.join(cfg_dir, 'anomaly_detection.txt')
+    command = f'bash -c "rm {adt_txt}"'
+    run_command(command)
+        
     if not success:
         logging.info("===========================================================================")
         logging.error(mismatches_exit_red)